requirements.txt
pyyaml
atlassian-python-api
selenium
